+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -3,9 +3,9 @@ package woowacourse.shopping
 import android.app.Application
 import woowacourse.shopping.data.DefaultCartRepository
 import woowacourse.shopping.data.DefaultProductRepository
-import woowacourse.shopping.di.RepositoryContainer
-import woowacourse.shopping.repository.CartRepository
-import woowacourse.shopping.repository.ProductRepository
+import woowacourse.shopping.data.repository.CartRepository
+import woowacourse.shopping.data.repository.ProductRepository
+import woowacourse.shopping.di.DIContainer
 
 class ShoppingApplication : Application() {
     override fun onCreate() {
