+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data
 
+import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.repository.CartRepository
 
 // TODO: Step2 - CartProductDao를 참조하도록 변경
 class DefaultCartRepository : CartRepository {
