+++ app/src/main/java/woowacourse/shopping/data/DefaultProductRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data
 
+import woowacourse.shopping.data.repository.ProductRepository
 import woowacourse.shopping.model.Product
-import woowacourse.shopping.repository.ProductRepository
 
 class DefaultProductRepository : ProductRepository {
     private val products: List<Product> =
